ingredient feauture

+++ controller/ingredients.controller.ts
@@ -1,8 +1,10 @@
 import { AuthRequest } from "../types/request.type";
 import { Request, Response } from "express";
-import { createIngredients } from "../service/ingredient.service";
+import { createIngredients , getIngredientsByBranch} from "../service/ingredient.service";
 import fs from "fs";
 import cloudinary from "../utils/cloudinary"
+import { findAccountById } from "../service/account.service";
+import { accountInterface } from "../types/account.type";
 
 
 
@@ -12,14 +14,32 @@ export const createIngredientsController = async (request: AuthRequest, response
             response.status(400).json({ error: 'No file uploaded' });
             return;
         }
+        
+        if(!request.id){
+            console.log("not autenitcated")
+            response.status(500).json({ error: 'not autenitcated' });
+            return;
+        }
 
         const uploadResult = await cloudinary.uploader.upload(request.file.path, {
         folder: 'nextjs_uploads',
         });
 
-        fs.unlinkSync(request.file.path); 
+        fs.unlinkSync(request.file.path);
 
-        response.json({ url: uploadResult.secure_url });
+        const url = uploadResult.secure_url
+        const { name, stocks } = request.body;
+
+        const account = await findAccountById(request.id)
+
+        if(!account){
+            response.status(404).json({ error: 'Account not found' });
+            return;
+        }
+   
+
+        await createIngredients({ name, stocks, branch: account.branch, img: url });
+        response.send("successfully uploaded");
     } catch (error) {
         console.error(error);
         response.status(500).json({ error: 'Upload failed' });
@@ -28,3 +48,16 @@ export const createIngredientsController = async (request: AuthRequest, response
     
 };
 
+
+export const getIngredientsController = async (request: AuthRequest, response: Response) => {
+    if(!request.id)
+    {
+        response.status(500).send("not authenticated")
+        return
+    }
+
+    const account = await findAccountById(request.id);
+
+    const ingredients = await getIngredientsByBranch(account?.branch || "");
+    response.send(ingredients);
+}
